# docs/swagger.yaml
definitions:
...
  disability_system_backend_internal_modules_incapacidades_dto.ActualizarIncapacidadRequest:
    properties:
      canal_recepcion:
        type: string
      fecha_fin:
        type: string
      fecha_inicio:
        type: string
      fecha_pago:
        type: string
      fecha_radicacion:
        type: string
      id_entidad:
        type: integer
      id_tipo:
        type: integer
      id_usuario:
        type: integer
      observaciones:
        type: string
      origen:
        type: string
      titulo:
        type: string
    type: object
  disability_system_backend_internal_modules_incapacidades_dto.CambiarEstadoRequest:
    properties:
      id_estado:
        type: integer
      observaciones:
        type: string
    required:
    - id_estado
    type: object
  disability_system_backend_internal_modules_incapacidades_dto.CrearIncapacidadRequest:
    properties:
      canal_recepcion:
        type: string
      fecha_fin:
        type: string
      fecha_inicio:
        type: string
      fecha_pago:
        type: string
      fecha_radicacion:
        type: string
      id_entidad:
        type: integer
      id_estado:
        type: integer
      id_tipo:
        type: integer
      id_usuario:
        type: integer
      observaciones:
        type: string
      origen:
        type: string
      titulo:
        type: string
    required:
    - fecha_inicio
    - id_entidad
    - id_tipo
    - origen
    - titulo
    type: object
...
info:
  contact: {}
paths:
...
  /catalogos/estados-documento:
    get:
      consumes:
      - application/json
      description: Lista todos los estados de documento
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar estados de documento
      tags:
      - catalogos
  /catalogos/tipos-documento:
    get:
      consumes:
      - application/json
      description: Lista todos los tipos de documento
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar tipos de documento
      tags:
      - catalogos
  /catalogos/tipos-pago:
    get:
      consumes:
      - application/json
      description: Lista todos los tipos de pago
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar tipos de pago
      tags:
      - catalogos
...
  /incapacidades:
    get:
      consumes:
      - application/json
      description: Lista las incapacidades con filtros y paginación
      parameters:
      - description: Filtrar por usuario
        in: query
        name: id_usuario
        type: integer
      - description: Filtrar por estado
        in: query
        name: id_estado
        type: integer
      - description: Filtrar por tipo
        in: query
        name: id_tipo
        type: integer
      - description: Filtrar por entidad
        in: query
        name: id_entidad
        type: integer
      - description: Filtrar por origen
        in: query
        name: origen
        type: string
      - description: Filtrar por canal
        in: query
        name: canal_recepcion
        type: string
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar incapacidades
      tags:
      - incapacidades
    post:
      consumes:
      - application/json
      description: Registra una nueva incapacidad médica
      parameters:
      - description: Datos de la incapacidad
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_incapacidades_dto.CrearIncapacidadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear incapacidad
      tags:
      - incapacidades
  /incapacidades/{id}:
    delete:
      consumes:
      - application/json
      description: Archiva una incapacidad (soft delete)
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Archivar incapacidad
      tags:
      - incapacidades
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de una incapacidad específica
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener incapacidad por ID
      tags:
      - incapacidades
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una incapacidad
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_incapacidades_dto.ActualizarIncapacidadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar incapacidad
      tags:
      - incapacidades
...
  /incapacidades/{id}/estado:
    patch:
      consumes:
      - application/json
      description: Actualiza el estado de una incapacidad
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      - description: Nuevo estado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_incapacidades_dto.CambiarEstadoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cambiar estado de incapacidad
      tags:
      - incapacidades
...
  /incapacidades/{id}/plazos:
    get:
      consumes:
      - application/json
      description: Obtiene los plazos y alertas de vencimiento para una incapacidad
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener información de plazos
      tags:
      - incapacidades
...
  /incapacidades/entidades:
    get:
      consumes:
      - application/json
      description: Lista todas las EPS y ARL registradas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar entidades
      tags:
      - catalogos
  /incapacidades/estados:
    get:
      consumes:
      - application/json
      description: Lista todos los estados disponibles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar estados de incapacidad
      tags:
      - catalogos
  /incapacidades/tipos:
    get:
      consumes:
      - application/json
      description: Lista todos los tipos de incapacidad
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar tipos de incapacidad
      tags:
      - catalogos
  /incapacidades/tipos/{tipo_id}/documentos-requeridos:
    get:
      consumes:
      - application/json
      description: Lista los documentos requeridos para un tipo de incapacidad
      parameters:
      - description: ID del tipo de incapacidad
        in: path
        name: tipo_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener documentos requeridos
      tags:
      - incapacidades
...
swagger: "2.0"